added/relocated a strategyWebContentProvider
